working on git pull comments; fixes from testing
